requirements.txt
behave
requests
allure-behave
python-dotenv
PyHamcrest